contract: logic_start_mtc merge into storage
