+++ contract/logic_start_mtc/lib.rs
@@ -30,77 +30,6 @@ pub mod contract {
             }
         }
 
-        #[ink(message)]
-        pub fn start_mtc(&mut self, caller: AccountId, deck_emo_base_ids: [u16; 6]) {
-            self.only_allowed_caller();
-
-            let mut storage = StorageRef::from_account_id(self.storage_account_id);
-
-            let (
-                emo_bases,
-                deck_fixed_emo_base_ids,
-                deck_built_emo_base_ids,
-                player_ep,
-                _,
-                _,
-                player_health,
-                ..,
-            ) = storage.get_player_batch(
-                caller, true, true, true, true, false, false, true, false, false, false, false,
-                false,
-            );
-
-            let ep = if player_health.is_some() {
-                // the previous mtc didn't normally finish
-                player_ep
-                    .expect("player ep none")
-                    .saturating_sub(EP_UNFINISH_PENALTY)
-            } else {
-                player_ep.unwrap_or(INITIAL_EP)
-            };
-
-            let seed = self.get_random_seed(caller, b"start_mtc");
-
-            let selected_ghosts =
-                choose_ghosts(ep, seed, &|ep_band| storage.get_matchmaking_ghosts(ep_band));
-
-            storage.set_player_batch(
-                caller,
-                if let Some(e) = player_ep {
-                    if e == ep {
-                        None
-                    } else {
-                        Some(ep)
-                    }
-                } else {
-                    Some(ep)
-                },
-                Some(seed),
-                Some(
-                    build_pool(
-                        &deck_emo_base_ids,
-                        &emo_bases.expect("emo_bases none"),
-                        &deck_fixed_emo_base_ids.expect("deck_fixed_emo_base_ids none"),
-                        &deck_built_emo_base_ids.expect("deck_built_emo_base_ids none"),
-                    )
-                    .expect("failed to build player pool"),
-                ),
-                Some(PLAYER_INITIAL_HEALTH),
-                Some(Vec::new()),
-                Some(get_upgrade_coin(2)),
-                Some(selected_ghosts),
-                Some(build_initial_ghost_states()),
-                Some(0),
-            );
-        }
-
-        fn get_random_seed(&self, caller: AccountId, subject: &[u8]) -> u64 {
-            let (seed, _) = self
-                .env()
-                .random(&self.env().hash_encoded::<Blake2x128, _>(&(subject, caller)));
-            <u64>::decode(&mut seed.as_ref()).expect("failed to get seed")
-        }
-
         // allowed accounts
 
         #[ink(message)]
